mvp4/agent_testing/requirement.txt
elevenlabs==1.8.0
google-generativeai==0.7.2
sounddevice==0.4.6
scipy==1.11.4
numpy==1.26.4
pyaudio
pydub==0.25.1
playsound==1.3.0
requests==2.31.0

fastapi
uvicorn[standard]
python-dotenv
google-generativeai
supabase
langgraph
langchain-core
tzdata


elevenlabs==1.8.0
google-generativeai==0.7.2
sounddevice==0.4.6
scipy==1.11.4
numpy==1.26.4
pyaudio
pydub==0.25.1
playsound==1.3.0
requests==2.31.0
